Expose TTS profile to clients
This change allows clients to access the `profile` argument
for the GraphQL synthesis method so that audio formatted for
third-party services can be produced. It also condenses the
GraphQL query creation and adds some documentation for the
new option.

Default values have been added to the manager's `synthesize`
method to support some example code from the readme. The
enclosing GraphQL query has also been changed to include the
platform name.